requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
-e git+git://github.com/chauffer/cobe.git#egg=cobe  # via -r requirements.in
ago==0.0.93               # via -r requirements.in
aiogoogletrans==3.3.2     # via -r requirements.in
aiohttp==3.6.2            # via -r requirements.in, aiogoogletrans, discord.py
aioredis==1.3.1           # via -r requirements.in
aioyoutube==0.1.6         # via -r requirements.in
aniso8601==8.0.0          # via -r requirements.in
argparse==1.4.0           # via cobe
arrow==0.14.7             # via -r requirements.in, ics
async-timeout==3.0.1      # via aiohttp, aioredis
asyncio==3.4.3            # via -r requirements.in
attrs==19.3.0             # via aiohttp
cffi==1.14.1              # via pynacl
chardet==3.0.4            # via aiohttp
colorama==0.4.3           # via -r requirements.in
discord.py[voice]==1.3.4  # via -r requirements.in
fuzzywuzzy[speedup]==0.18.0  # via -r requirements.in
hiredis==1.1.0            # via aioredis
humanize==2.5.0           # via -r requirements.in
ics==0.7                  # via -r requirements.in
idna==2.10                # via yarl
multidict==4.7.6          # via aiohttp, yarl
pycparser==2.20           # via cffi
pynacl==1.3.0             # via discord.py
git+git://github.com/snowballstem/pystemmer@master#egg=PyStemmer  # via cobe
python-dateutil==2.8.1    # via arrow, ics
python-levenshtein==0.12.0  # via fuzzywuzzy
rfc3339==6.2              # via aioyoutube
six==1.15.0               # via ics, pynacl, python-dateutil
structlog==20.1.0         # via -r requirements.in
tatsu==5.5.0              # via ics
typing-extensions==3.7.4.2  # via yarl
websockets==8.1           # via discord.py
yarl==1.5.0               # via aiohttp
youtube-dl==2020.7.28     # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
